refactor(modules): Split AppModule into separate modules
- dashboard-last-values moved to separate module
- a lot of related dialogs also moved to dashboard-last-values.module.ts
