+++ src/app/dashboard/dashboard-last-values/dialogs/saved-gas-fees-dialog/saved-gas-fees-dialog.component.ts
@@ -1,0 +1,45 @@
+import {AfterViewInit, ChangeDetectorRef, Component} from '@angular/core';
+import {HttpService} from '../../../../services/http/http.service';
+import {ViewTypeService} from '../../../../services/view-type.service';
+import {NGXLogger} from 'ngx-logger';
+import {ChartBuilder} from '../../../../chart/chart-builder';
+import {ChartGeneralMethodsComponent} from 'src/app/chart/chart-general-methods.component';
+import {HardworksService} from '../../../../services/http/hardworks.service';
+import {StaticValues} from '../../../../static/static-values';
+
+@Component({
+  selector: 'app-saved-gas-fees-dialog',
+  templateUrl: './saved-gas-fees-dialog.component.html',
+  styleUrls: ['./saved-gas-fees-dialog.component.scss']
+})
+export class SavedGasFeesDialogComponent extends ChartGeneralMethodsComponent implements AfterViewInit {
+
+  constructor(private httpService: HttpService,
+              public vt: ViewTypeService,
+              public cdRef: ChangeDetectorRef,
+              private log: NGXLogger,
+              private hardworksService: HardworksService,
+  ) {
+    super(cdRef, vt);
+  }
+
+  load(): void {
+    this.hardworksService.getHardWorkHistoryData(StaticValues.NETWORKS.get(this.network))
+      .subscribe((data) => {
+        this.log.debug('History of All Hard Works loaded ', data);
+        const chartBuilder = new ChartBuilder();
+        const hwFees = new Map<number, number>();
+        let savedGas = 0;
+
+        data?.forEach(dto => {
+          savedGas += dto.savedGasFees;
+          hwFees.set(dto.blockDate, savedGas);
+        });
+        chartBuilder.initVariables(1);
+        hwFees.forEach((fees, date) => chartBuilder.addInData(0, date, fees / 1000000));
+        this.handleData(chartBuilder, [
+          ['Saved Gas Fees M$', 'right', '#0085ff']
+        ]);
+      });
+  }
+}
