+++ src/app/dashboard/dashboard-last-values/dialogs/tvl-v2-dialog/tvl-v2-dialog.component.scss
