+++ src/app/dashboard/dashboard-last-values/dialogs/weekly-profit-dialog/weekly-profit-dialog.component.spec.ts
